Extract search out from seeds api directory
- now that we use it for more than just search
- tested across webpage
