+++ src/components/seeds/checkin/index.tsx
@@ -14,7 +14,8 @@ import { makeStyles, withStyles } from '@mui/styles';
 import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router';
 import { useRecoilState } from 'recoil';
-import { getPendingAccessions, SearchResponseElement } from 'src/api/seeds/search';
+import { getPendingAccessions } from 'src/api/seeds/search';
+import { SearchResponseElement } from 'src/api/search';
 import Button from 'src/components/common/button/Button';
 import TfMain from 'src/components/common/TfMain';
 import { APP_PATHS } from 'src/constants';
