+++ src/components/seeds/database/filters/FilterCountWeight.tsx
@@ -3,7 +3,7 @@ import { Grid } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import { Theme } from '@mui/material';
 import React from 'react';
-import { AndNodePayload, FieldNodePayload, OrNodePayload, SearchNodePayload } from 'src/api/seeds/search';
+import { AndNodePayload, FieldNodePayload, OrNodePayload, SearchNodePayload } from 'src/api/search';
 import Checkbox from 'src/components/common/Checkbox';
 import Divisor from 'src/components/common/Divisor';
 import Dropdown from 'src/components/common/Dropdown';
