+++ src/components/seeds/database/filters/FilterDateRange.tsx
@@ -3,7 +3,7 @@ import { Grid, Theme } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import { format, isValid } from 'date-fns';
 import React from 'react';
-import { FieldNodePayload } from 'src/api/seeds/search';
+import { FieldNodePayload } from 'src/api/search';
 import DatePicker from 'src/components/common/DatePicker';
 import strings from 'src/strings';
