+++ src/components/seeds/database/filters/FilterNumberRange.tsx
@@ -2,7 +2,7 @@ import { Grid, Theme } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import React from 'react';
 import TextField from '../../../common/TextField';
-import { FieldNodePayload } from '../../../../api/seeds/search';
+import { FieldNodePayload } from '../../../../api/search';
 import { ArrowForward } from '@mui/icons-material';
 
 const useStyles = makeStyles((theme: Theme) => ({
