+++ src/components/seeds/database/filters/FilterSearch.tsx
@@ -2,7 +2,7 @@ import { Search as SearchIcon } from '@mui/icons-material';
 import { InputAdornment, TextField, Theme } from '@mui/material';
 import React from 'react';
 import strings from 'src/strings';
-import { FieldNodePayload } from '../../../../api/seeds/search';
+import { FieldNodePayload } from '../../../../api/search';
 import { makeStyles } from '@mui/styles';
 
 const useStyles = makeStyles((theme: Theme) => ({
